requirements.txt
Flask==2.0.1
Flask-Uploads==0.2.1
python-pptx==0.6.19
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.21.0
requests==2.25.1

